refactored handles module, created two speculation modules: armoire (CADHT) and estuary (VCSL)
