+++ src/armoire/mod.rs
@@ -1,0 +1,1 @@
+// Content addressed hash table
