+++ src/handle/mod.rs
@@ -1,9 +1,13 @@
-pub mod tag;
 pub mod address;
 pub mod identity;
 pub mod location;
 
-pub use tag::Tag;
 pub use address::Address;
 pub use identity::Identity;
 pub use location::Location;
+
+pub fn hex(bytes: &[u8]) -> String {
+    bytes.iter()
+        .map(|b| format!("{:02x}", b))
+        .collect::<String>()
+}
